+++ udata/event/values.py
@@ -5,12 +5,6 @@ class KafkaMessageType(Enum):
     INDEX = 'index'
     REINDEX = 'reindex'
     UNINDEX = 'unindex'
-    RESOURCE_CREATED = 'resource_created'
-    RESOURCE_MODIFIED = 'resource_modified'
-    RESOURCE_DELETED = 'resource_removed'
-
-
-class KafkaTopic(Enum):
-    RESOURCE_CREATED = 'resource.created'
-    RESOURCE_MODIFIED = 'resource.modified'
-    RESOURCE_DELETED = 'resource.removed'
+    CREATED = 'created'
+    MODIFIED = 'modified'
+    DELETED = 'removed'